+++ fjs/types/rtti/proof.f.ts
@@ -1,6 +1,6 @@
 import type { StringMap } from '../object/module.f.ts'
 
-type Tests = StringMap<string, readonly unknown[]>
+type Tests = StringMap<readonly unknown[]>
 
 const tests: Tests = {
     undefined: [undefined],